# refactor: analyze css dependencies with css-module-lexer