--- a/crates/rspack_core/src/dependency/dependency_category.rs
+++ b/crates/rspack_core/src/dependency/dependency_category.rs
@@ -10,8 +10,8 @@ pub enum DependencyCategory {
   Url,
   CssImport,
   CssCompose,
-  /// FIXME: remove after we align css module with webpack
-  CssModuleExport,
+  CssExport,
+  CssLocalIdent,
   Wasm,
   Worker,
   LoaderImport,
@@ -25,8 +25,9 @@ impl From<&str> for DependencyCategory {
       "url" => Self::Url,
       "wasm" => Self::Wasm,
       "css-import" => Self::CssImport,
-      "css-export" => Self::CssModuleExport,
+      "css-export" => Self::CssExport,
       "css-compose" => Self::CssCompose,
+      "css-local-ident" => Self::CssLocalIdent,
       "worker" => Self::Worker,
       "unknown" => Self::Unknown,
       _ => unimplemented!("DependencyCategory {}", value),
@@ -43,10 +44,11 @@ impl DependencyCategory {
       DependencyCategory::Url => "url",
       DependencyCategory::CssImport => "css-import",
       DependencyCategory::CssCompose => "css-compose",
+      DependencyCategory::CssExport => "css-export",
+      DependencyCategory::CssLocalIdent => "css-local-ident",
       DependencyCategory::Wasm => "wasm",
       DependencyCategory::Worker => "worker",
       DependencyCategory::LoaderImport => "loader import",
-      DependencyCategory::CssModuleExport => "css-export",
     }
   }
 }