--- a/crates/rspack_core/src/dependency/dependency_type.rs
+++ b/crates/rspack_core/src/dependency/dependency_type.rs
@@ -52,9 +52,10 @@ pub enum DependencyType {
   CssImport,
   // css modules compose
   CssCompose,
-  /// css module export
-  /// FIXME: remove after we align css module with webpack
-  CssModuleExport,
+  // css :export
+  CssExport,
+  // css modules local ident
+  CssLocalIdent,
   // context element
   ContextElement,
   // import context
@@ -124,7 +125,8 @@ impl DependencyType {
       DependencyType::CssUrl => Cow::Borrowed("css url"),
       DependencyType::CssImport => Cow::Borrowed("css import"),
       DependencyType::CssCompose => Cow::Borrowed("css compose"),
-      DependencyType::CssModuleExport => Cow::Borrowed("css export"),
+      DependencyType::CssExport => Cow::Borrowed("css export"),
+      DependencyType::CssLocalIdent => Cow::Borrowed("css local ident"),
       DependencyType::ContextElement => Cow::Borrowed("context element"),
       // TODO: mode
       DependencyType::ImportContext => Cow::Borrowed("import context"),