--- a/crates/rspack_plugin_css/src/dependency/mod.rs
+++ b/crates/rspack_plugin_css/src/dependency/mod.rs
@@ -1,9 +1,11 @@
 mod compose;
-mod css_module_export;
+mod export;
 mod import;
+mod local_ident;
 mod url;
 
 pub use compose::*;
-pub use css_module_export::*;
+pub use export::*;
 pub use import::*;
+pub use local_ident::*;
 pub use url::*;