--- a/crates/rspack_plugin_css/src/lib.rs
+++ b/crates/rspack_plugin_css/src/lib.rs
@@ -5,8 +5,6 @@ pub mod dependency;
 mod parser_and_generator;
 pub mod plugin;
 pub mod runtime;
-pub mod swc_css_compiler;
 mod utils;
-pub mod visitors;
 
 pub use plugin::CssPlugin;