--- a/crates/rspack_plugin_css/src/lib.rs
+++ b/crates/rspack_plugin_css/src/lib.rs
@@ -1,6 +1,8 @@
 #![feature(let_chains)]
 #![feature(if_let_guard)]
 #![feature(box_patterns)]
+#![feature(option_get_or_insert_default)]
+
 pub mod dependency;
 mod parser_and_generator;
 pub mod plugin;