--- a/crates/rspack_plugin_css/src/plugin/impl_plugin_for_css_plugin.rs
+++ b/crates/rspack_plugin_css/src/plugin/impl_plugin_for_css_plugin.rs
@@ -355,7 +355,7 @@ impl Plugin for CssPlugin {
           .and_then(|g| g.get_css_module())
           .expect("should have CssModuleGeneratorOptions");
         Box::new(CssParserAndGenerator {
-          exports: None,
+          exports: Some(Default::default()),
           convention: Some(
             g.exports_convention
               .expect("should have exports_convention"),
@@ -381,7 +381,7 @@ impl Plugin for CssPlugin {
           .and_then(|g| g.get_css_auto())
           .expect("should have CssAutoGeneratorOptions");
         Box::new(CssParserAndGenerator {
-          exports: None,
+          exports: Some(Default::default()),
           convention: Some(
             g.exports_convention
               .expect("should have exports_convention"),