--- a/crates/rspack_plugin_javascript/src/parser_and_generator/mod.rs
+++ b/crates/rspack_plugin_javascript/src/parser_and_generator/mod.rs
@@ -321,6 +321,7 @@ impl ParserAndGenerator for JavaScriptParserAndGenerator {
         init_fragments: &mut init_fragments,
         runtime: generate_context.runtime,
         concatenation_scope: generate_context.concatenation_scope.take(),
+        data: generate_context.data,
       };
 
       module.get_dependencies().iter().for_each(|dependency_id| {