--- a/webpack-examples/loader/webpack.config.js
+++ b/webpack-examples/loader/webpack.config.js
@@ -7,8 +7,5 @@ module.exports = {
 				loader: "css-loader"
 			}
 		]
-	},
-	experiments: {
-		css: false,
 	}
 };