… Cargo.lock
… Cargo.toml
… crates/node_binding/Cargo.toml
… crates/rspack_plugin_css/Cargo.toml
… crates/rspack_plugin_css/src/utils.rs
… crates/rspack_plugin_lazy_compilation/src/module.rs
… crates/rspack_plugin_swc_css_minimizer/Cargo.toml
… crates/rspack_plugin_swc_css_minimizer/src/lib.rs
… packages/rspack-cli/tests/build/basic/public/main.js
… packages/rspack-cli/tests/build/basic/public/main.js.map
… packages/rspack-test-tools/tests/__snapshots__/Config.test.js.snap
… packages/rspack-test-tools/tests/builtinCases/plugin-css-modules/modules-composes/__snapshots__/output.snap.txt
… packages/rspack-test-tools/tests/builtinCases/plugin-css-modules/modules-ident-name/__snapshots__/output.snap.txt
… packages/rspack-test-tools/tests/builtinCases/plugin-css/at-import-in-the-middle/__snapshots__/output.snap.txt
… packages/rspack-test-tools/tests/diagnosticsCases/factorize/css-module-request-prefix/index.css
… packages/rspack-test-tools/tests/diagnosticsCases/factorize/css-module-request-prefix/stats.err
… packages/rspack-test-tools/tests/diagnosticsCases/module-parse-failed/at-import-in-the-middle/a.css
… packages/rspack-test-tools/tests/diagnosticsCases/module-parse-failed/at-import-in-the-middle/b.css
… packages/rspack-test-tools/tests/diagnosticsCases/module-parse-failed/at-import-in-the-middle/c.css
… packages/rspack-test-tools/tests/diagnosticsCases/module-parse-failed/at-import-in-the-middle/index.css
… packages/rspack-test-tools/tests/diagnosticsCases/module-parse-failed/at-import-in-the-middle/index.js
… packages/rspack-test-tools/tests/diagnosticsCases/module-parse-failed/at-import-in-the-middle/stats.err