<!--
  Thank you for submitting a pull request!

  We appreciate the time and effort you have invested in making these changes. Please ensure that you provide enough information to allow others to review your pull request.

  Upon submission, your pull request will be automatically assigned with reviewers.

  If you want to learn more about contributing to this project, please visit: https://github.com/web-infra-dev/rspack/blob/main/CONTRIBUTING.md.
-->

## Summary

Switch css parser from `swc-css` to `css-module-lexer`, solve not correctly finding css modules local class issue

for example:

```less
// style.less
.wrapper {
  display: flex;
  &:global(.semi) {
    &.cancel_padding_right {
      padding-right: 0;
    }
  }
}
```

before:

```css
.__src_b_wrapper {
  display: flex;
}
.__src_b_wrapper.semi {
  padding-right: 0;
}
```

after:

```css
.__src_b_wrapper {
  display: flex;
}
.__src_b_wrapper.semi.__src_b_cancel_padding_right {
  padding-right: 0;
}
```

And ~4x faster for bundling `import "bootstrap/dist/css/bootstrap.css"`

before: ~84ms (~71ms for analyzing css dependencies)

after: ~25ms (~11ms for analyzing css dependencies)

<!-- Can you explain the reasoning behind implementing this change? What problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [ ] Documentation updated (or not required).